Moves method participants into separate class
