--- src/main/scala/scala/tools/refactoring/implementations/ClassDefOrganizeImports.scala
@@ -1,8 +1,0 @@
-package scala.tools.refactoring
-package implementations
-
-trait ClassDefOrganizeImports { self: OrganizeImports =>
-  import global._
-
-  def organizeImportsInClassDefs(tree: Tree): Tree = tree
-}
